+++ kernel/bio.c
@@ -23,69 +23,236 @@
 #include "fs.h"
 #include "buf.h"
 
-struct {
-  struct spinlock lock;
-  struct buf buf[NBUF];
+/**
+ * @brief 改进了bcache。
+ *
+ * 增加了哈希表；
+ * 移除了链表；
+ * 增加了堆。
+ */
+struct
+{
+    struct spinlock lock;
+    struct buf buf[NBUF];
 
-  // Linked list of all buffers, through prev/next.
-  // Sorted by how recently the buffer was used.
-  // head.next is most recent, head.prev is least.
-  struct buf head;
+    /**
+     * @brief 保存buffer指针的哈希表。
+     */
+    struct buf *hash[NBUF];
+
+    /**
+     * @brief 保存buffer指针的堆。
+     */
+    struct buf *heap[NBUF];
+
+    /**
+     * @brief 目前空闲的buffer数量。
+     */
+    uint heapSize;
+
+    /**
+     * @brief 用于LRU算法的时间戳。
+     *
+     * 每次访问buffer，该时间戳+1。
+     */
+    uint timeStamp;
 } bcache;
 
-void
-binit(void)
+/**
+ * @brief 改进了binit。
+ *
+ * 增加了哈希表的初始化；
+ * 移除了链表的初始化；
+ * 增加了堆的初始化。
+ */
+void binit()
 {
-  struct buf *b;
+    struct buf *b;
 
-  initlock(&bcache.lock, "bcache");
+    initlock(&bcache.lock, "bcache");
+    bcache.heapSize = 0;
 
-  // Create linked list of buffers
-  bcache.head.prev = &bcache.head;
-  bcache.head.next = &bcache.head;
-  for(b = bcache.buf; b < bcache.buf+NBUF; b++){
-    b->next = bcache.head.next;
-    b->prev = &bcache.head;
-    initsleeplock(&b->lock, "buffer");
-    bcache.head.next->prev = b;
-    bcache.head.next = b;
-  }
+    for (int i = 0; i < NBUF; i++)
+    {
+        b = bcache.buf + i;
+        initsleeplock(&b->lock, "buffer");
+
+        // 初始状态，所有buffer均不在哈希表中，且在堆中
+        b->prev = 0;
+        b->next = 0;
+        bcache.hash[i] = 0;
+
+        b->timeStamp = 0;
+        b->heapIndex = i;
+        bcache.heap[i] = b;
+    }
 }
 
-// Look through buffer cache for block on device dev.
-// If not found, allocate a buffer.
-// In either case, return locked buffer.
-static struct buf*
-bget(uint dev, uint blockno)
+/**
+ * @brief 在哈希表中寻找指定buffer
+ *
+ * @return 如果找到，返回该buffer的地址；
+ * 如果未找到，返回空指针。
+ */
+static struct buf *bFindFromHashTable(uint dev, uint blockno)
 {
-  struct buf *b;
+    uint hash = (dev + blockno) % NBUF;
 
-  acquire(&bcache.lock);
-
-  // Is the block already cached?
-  for(b = bcache.head.next; b != &bcache.head; b = b->next){
-    if(b->dev == dev && b->blockno == blockno){
-      b->refcnt++;
-      release(&bcache.lock);
-      acquiresleep(&b->lock);
-      return b;
+    if (bcache.hash[hash] == 0)
+        return 0;
+    else
+    {
+        struct buf *head = bcache.hash[hash];
+        while (head != 0)
+        {
+            if (head->dev == dev && head->blockno == blockno)
+                return head;
+            head = head->next;
+        }
     }
-  }
 
-  // Not cached.
-  // Recycle the least recently used (LRU) unused buffer.
-  for(b = bcache.head.prev; b != &bcache.head; b = b->prev){
-    if(b->refcnt == 0) {
-      b->dev = dev;
-      b->blockno = blockno;
-      b->valid = 0;
-      b->refcnt = 1;
-      release(&bcache.lock);
-      acquiresleep(&b->lock);
-      return b;
+    return 0;
+}
+
+/**
+ * @brief 从指定位置开始上滤
+ */
+static void bPercolateUp(int index)
+{
+    struct buf *value = bcache.heap[index];
+    int parentIndex = (index - 1) / 2;
+
+    while (index >= 0 && bcache.heap[parentIndex]->timeStamp > value->timeStamp)
+    {
+        bcache.heap[index] = bcache.heap[parentIndex];
+        bcache.heap[index]->heapIndex = index;
+        index = parentIndex;
+        parentIndex = (index - 1) / 2;
     }
-  }
-  panic("bget: no buffers");
+
+    bcache.heap[index] = value;
+    value->heapIndex = index;
+}
+
+/**
+ * @brief 从指定位置开始下滤
+ */
+static void bPercolateDown(int index)
+{
+    int maxChild = 2 * (index + 1);
+    struct buf *value = bcache.heap[index];
+    char goDown = 1;
+    while (goDown && maxChild < bcache.heapSize)
+    {
+        goDown = 0;
+        if (bcache.heap[maxChild]->timeStamp < bcache.heap[maxChild - 1]->timeStamp)
+            --maxChild;
+        if (value->timeStamp < bcache.heap[maxChild]->timeStamp)
+        {
+            goDown = 1;
+            bcache.heap[index] = bcache.heap[maxChild];
+            bcache.heap[index]->heapIndex = index;
+            index = maxChild;
+            maxChild = 2 * (maxChild + 1);
+        }
+    }
+    if (maxChild == bcache.heapSize)
+    {
+        if (value->timeStamp < bcache.heap[maxChild - 1]->timeStamp)
+        {
+            bcache.heap[index] = bcache.heap[maxChild - 1];
+            bcache.heap[index]->heapIndex = index;
+            index = maxChild - 1;
+        }
+    }
+    bcache.heap[index] = value;
+    value->heapIndex = index;
+}
+
+/**
+ * @brief 改进了bget。
+ *
+ * 获取已缓存buffer的过程改为由哈希表获取；
+ * 分配新buffer后将其插入哈希表。
+ */
+static struct buf *bget(uint dev, uint blockno)
+{
+    struct buf *b;
+    struct buf **index;
+    uint hash;
+
+    acquire(&bcache.lock);
+
+    b = bFindFromHashTable(dev, blockno);
+
+    if (b != 0)
+    {
+        b->refcnt++;
+
+        if (b->refcnt == 1)
+        {
+            //从堆中移除b
+            int bIndex = b->heapIndex;
+            bcache.heap[bIndex] = bcache.heap[bcache.heapSize - 1];
+            bcache.heap[bcache.heapSize - 1] = b;
+            bcache.heapSize--;
+            b->heapIndex = NBUF;
+            if (bcache.heap[bIndex]->timeStamp < bcache.heap[(bIndex - 1) / 2])
+                bPercolateUp(bIndex);
+            if (bcache.heap[bIndex]->timeStamp > bcache.heap[2 * (bIndex + 1)]->timeStamp)
+                bPercolateDown(bIndex);
+            if (bcache.heap[bIndex]->timeStamp > bcache.heap[2 * bIndex + 1]->timeStamp)
+                bPercolateDown(bIndex);
+        }
+
+        release(&bcache.lock);
+        acquiresleep(&b->lock);
+        return b;
+    }
+
+    while (bcache.heapSize > 0)
+    {
+        //从堆中取出一个空闲buffer，相当于LRU算法
+        b = *(bcache.heap);
+        bcache.heapSize--;
+        *(bcache.heap) = bcache.heap[bcache.heapSize];
+        bPercolateDown(0);
+
+        if (b->refcnt == 0)
+        {
+            //从哈希表中移除b
+            if(b->prev == 0)
+            {
+                hash = (b->dev + b->blockno) % NBUF;
+                bcache.hash[hash] = b->next;
+            }
+            else
+                b->prev->next = b->next;
+
+            if(b->next != 0)
+                b->next->prev = b->prev;
+
+            //原bget函数中的相关操作
+            b->dev = dev;
+            b->blockno = blockno;
+            b->valid = 0;
+            b->refcnt = 1;
+            release(&bcache.lock);
+            acquiresleep(&b->lock);
+
+            //把新分配的buffer插入到哈希表的空位
+            hash = (b->dev + b->blockno) % NBUF;
+            b->next = bcache.hash[hash];
+            if(b->next != 0)
+                b->next->prev = b;
+            b->prev = 0;
+            bcache.hash[hash] = b;
+
+            return b;
+        }
+    }
+
+    panic("bget: no buffers");
 }
 
 // Return a locked buf with the contents of the indicated block.
@@ -111,31 +278,34 @@ bwrite(struct buf *b)
   virtio_disk_rw(b, 1);
 }
 
-// Release a locked buffer.
-// Move to the head of the most-recently-used list.
-void
-brelse(struct buf *b)
+/**
+ * @brief 改进了brelse。
+ *
+ * 增加了把buffer加入空闲buffer堆的步骤。
+ */
+void brelse(struct buf *b)
 {
-  if(!holdingsleep(&b->lock))
-    panic("brelse");
+    if (!holdingsleep(&b->lock))
+        panic("brelse");
 
-  releasesleep(&b->lock);
+    releasesleep(&b->lock);
 
-  acquire(&bcache.lock);
-  b->refcnt--;
-  if (b->refcnt == 0) {
-    // no one is waiting for it.
-    b->next->prev = b->prev;
-    b->prev->next = b->next;
-    b->next = bcache.head.next;
-    b->prev = &bcache.head;
-    bcache.head.next->prev = b;
-    bcache.head.next = b;
-  }
-  
-  release(&bcache.lock);
+    acquire(&bcache.lock);
+    b->refcnt--;
+    if (b->refcnt == 0)
+    {
+        //更新timeStamp
+        bcache.timeStamp++;
+        b->timeStamp = bcache.timeStamp;
+
+        //把b加入堆中
+        bcache.heap[bcache.heapSize] = b;
+        bcache.heapSize++;
+        bPercolateUp(bcache.heapSize - 1);
+    }
+
+    release(&bcache.lock);
 }
-
 void
 bpin(struct buf *b) {
   acquire(&bcache.lock);
