--- kernel/clm_param.h
@@ -1,20 +1,0 @@
-/**
- * @file clm_buf.h
- * @author Chlamydomonos
- * @brief 对param.h改进后的文件
- */
-
-#define NPROC        64  // maximum number of processes
-#define NCPU          8  // maximum number of CPUs
-#define NOFILE       16  // open files per process
-#define NFILE       100  // open files per system
-#define NINODE       50  // maximum number of active i-nodes
-#define NDEV         10  // maximum major device number
-#define ROOTDEV       1  // device number of file system root disk
-#define MAXARG       32  // max exec arguments
-#define MAXOPBLOCKS  10  // max # of blocks any FS op writes
-#define LOGSIZE      (MAXOPBLOCKS*3)  // max data blocks in on-disk log
-//把Buffer的数量调整到31个，便于计算哈希值
-#define NBUF         (MAXOPBLOCKS*3+1) // size of disk block cache
-#define FSSIZE       1000  // size of file system in blocks
-#define MAXPATH      128   // maximum file path name
