+++ kernel/param.h
@@ -8,6 +8,7 @@
 #define MAXARG       32  // max exec arguments
 #define MAXOPBLOCKS  10  // max # of blocks any FS op writes
 #define LOGSIZE      (MAXOPBLOCKS*3)  // max data blocks in on-disk log
-#define NBUF         (MAXOPBLOCKS*3)  // size of disk block cache
+//把Buffer的数量调整到31个，便于计算哈希值
+#define NBUF         (MAXOPBLOCKS*3+1) // size of disk block cache
 #define FSSIZE       1000  // size of file system in blocks
 #define MAXPATH      128   // maximum file path name
